Refatorações: aumentar coesão e redução de acoplamento
