+++ app/src/main/java/de/danoeh/antennapod/activity/MainActivity.java
@@ -1,20 +1,13 @@
 package de.danoeh.antennapod.activity;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.res.Configuration;
-import android.media.AudioManager;
-import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.RelativeLayout;
+
 import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
@@ -27,64 +20,40 @@ import androidx.core.view.WindowCompat;
 import androidx.core.view.WindowInsetsCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentContainerView;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.work.WorkInfo;
-import androidx.work.WorkManager;
+
 import com.bumptech.glide.Glide;
 import com.google.android.material.appbar.MaterialToolbar;
 import com.google.android.material.bottomsheet.BottomSheetBehavior;
 import com.google.android.material.snackbar.Snackbar;
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.event.EpisodeDownloadEvent;
-import de.danoeh.antennapod.event.FeedUpdateRunningEvent;
+import de.danoeh.antennapod.activity.handler.IntentHandler;
+import de.danoeh.antennapod.activity.handler.KeyboardInputHandler;
+import de.danoeh.antennapod.activity.manager.DrawerManager;
+import de.danoeh.antennapod.activity.manager.NavigationManager;
+import de.danoeh.antennapod.activity.manager.PlayerStateManager;
+import de.danoeh.antennapod.activity.observer.WorkManagerObserver;
 import de.danoeh.antennapod.event.MessageEvent;
-import de.danoeh.antennapod.model.download.DownloadStatus;
-import de.danoeh.antennapod.net.download.service.feed.FeedUpdateManagerImpl;
-import de.danoeh.antennapod.net.download.serviceinterface.DownloadServiceInterface;
-import de.danoeh.antennapod.net.download.serviceinterface.FeedUpdateManager;
-import de.danoeh.antennapod.net.sync.serviceinterface.SynchronizationQueue;
 import de.danoeh.antennapod.playback.cast.CastEnabledActivity;
-import de.danoeh.antennapod.playback.service.PlaybackServiceInterface;
-import de.danoeh.antennapod.storage.importexport.AutomaticDatabaseExportWorker;
-import de.danoeh.antennapod.storage.preferences.PlaybackPreferences;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
 import de.danoeh.antennapod.ui.TransitionEffect;
-import de.danoeh.antennapod.ui.appstartintent.MainActivityStarter;
-import de.danoeh.antennapod.ui.appstartintent.MediaButtonStarter;
-import de.danoeh.antennapod.ui.common.IntentUtils;
 import de.danoeh.antennapod.ui.common.ThemeSwitcher;
 import de.danoeh.antennapod.ui.common.ThemeUtils;
-import de.danoeh.antennapod.ui.discovery.DiscoveryFragment;
-import de.danoeh.antennapod.ui.screen.AddFeedFragment;
-import de.danoeh.antennapod.ui.screen.AllEpisodesFragment;
-import de.danoeh.antennapod.ui.screen.InboxFragment;
-import de.danoeh.antennapod.ui.screen.PlaybackHistoryFragment;
-import de.danoeh.antennapod.ui.screen.SearchFragment;
-import de.danoeh.antennapod.ui.screen.download.CompletedDownloadsFragment;
-import de.danoeh.antennapod.ui.screen.download.DownloadLogFragment;
 import de.danoeh.antennapod.ui.screen.drawer.BottomNavigation;
 import de.danoeh.antennapod.ui.screen.drawer.NavDrawerFragment;
 import de.danoeh.antennapod.ui.screen.drawer.NavigationNames;
-import de.danoeh.antennapod.ui.screen.feed.FeedItemlistFragment;
 import de.danoeh.antennapod.ui.screen.home.HomeFragment;
 import de.danoeh.antennapod.ui.screen.playback.audio.AudioPlayerFragment;
 import de.danoeh.antennapod.ui.screen.preferences.PreferenceActivity;
-import de.danoeh.antennapod.ui.screen.queue.QueueFragment;
 import de.danoeh.antennapod.ui.screen.rating.RatingDialogManager;
-import de.danoeh.antennapod.ui.screen.subscriptions.SubscriptionFragment;
 import de.danoeh.antennapod.ui.view.BottomSheetBackPressedCallback;
 import de.danoeh.antennapod.ui.view.LockableBottomSheetBehavior;
 import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.Validate;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class MainActivity extends CastEnabledActivity {
 
@@ -112,7 +81,7 @@ public class MainActivity extends CastEnabledActivity {
     private DrawerManager drawerManager;
     private PlayerStateManager playerStateManager;
     private NavigationManager navigationManager;
-    private StartupManager startupManager;
+    private KeyboardInputHandler.StartupManager startupManager;
     private IntentHandler intentHandler;
     private KeyboardInputHandler keyboardInputHandler;
     private WorkManagerObserver workManagerObserver;
@@ -135,7 +104,7 @@ public class MainActivity extends CastEnabledActivity {
         drawerManager = new DrawerManager(this, drawerLayout, navDrawer);
         playerStateManager = new PlayerStateManager(this);
         navigationManager = new NavigationManager(this);
-        startupManager = new StartupManager(this);
+        startupManager = new KeyboardInputHandler.StartupManager(this);
         intentHandler = new IntentHandler(this, navigationManager);
         keyboardInputHandler = new KeyboardInputHandler(this);
         workManagerObserver = new WorkManagerObserver(this);
