+++ app/src/main/java/de/danoeh/antennapod/activity/manager/NavigationManager.java
@@ -1,10 +1,11 @@
-package de.danoeh.antennapod.activity;
+package de.danoeh.antennapod.activity.manager;
 
 import android.os.Bundle;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import de.danoeh.antennapod.R;
+import de.danoeh.antennapod.activity.MainActivity;
 import de.danoeh.antennapod.ui.TransitionEffect;
 import de.danoeh.antennapod.ui.screen.AddFeedFragment;
 import de.danoeh.antennapod.ui.screen.AllEpisodesFragment;
